requirements.txt
altgraph==0.17.2; python_version >= "3.6" and python_version < "3.11" and sys_platform == "darwin"
appdirs==1.4.4
brotli==1.0.9; platform_python_implementation == "CPython" and python_version >= "3.6"
brotlicffi==1.0.9.2; platform_python_implementation == "PyPy" and python_version >= "3.6"
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
cffi==1.15.0; platform_python_implementation == "PyPy" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3"
colorama==0.4.4; python_version >= "3.5" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.5" and python_full_version >= "3.5.0"
cssutils==2.4.0; python_version >= "3.7"
future==0.18.2; python_version >= "3.6" and python_version < "3.11" and sys_platform == "win32" and python_full_version >= "3.6.0"
idna==3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
iniparse==0.5
loguru==0.6.0; python_version >= "3.5"
macholib==1.16; python_version >= "3.6" and python_version < "3.11" and sys_platform == "darwin"
multivolumefile==0.2.3; python_version >= "3.6"
packaging==21.3; python_version >= "3.6"
pefile==2021.9.3; python_full_version >= "3.6.0"
py7zr==0.18.1; python_version >= "3.6"
pybcj==0.5.0; platform_python_implementation == "CPython" and python_version >= "3.6"
pycparser==2.21; python_version >= "3.6" and python_full_version < "3.0.0" and platform_python_implementation == "PyPy" or platform_python_implementation == "PyPy" and python_version >= "3.6" and python_full_version >= "3.4.0"
pycryptodomex==3.14.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pyinstaller-hooks-contrib==2022.2; python_version >= "3.7" and python_version < "3.11"
pyinstaller==4.10; python_version >= "3.6" and python_version < "3.11"
pyparsing==3.0.7; python_version >= "3.6"
pyppmd==0.17.4; python_version >= "3.6"
pyside2==5.15.2.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "3.11"
pysimpleguiqt==0.35.0
pywin32-ctypes==0.2.0; python_version >= "3.6" and python_version < "3.11" and sys_platform == "win32"
pyzstd==0.15.2; python_version >= "3.6"
requests==2.27.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
ruamel.yaml.clib==0.2.6; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.5"
ruamel.yaml==0.17.21; python_version >= "3"
shiboken2==5.15.2.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "3.11"
six==1.16.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
texttable==1.6.4; python_version >= "3.6"
urllib3==1.26.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
win32-setctime==1.1.0; sys_platform == "win32" and python_version >= "3.5"
zipfile-deflate64==0.2.0; python_version >= "3.6"
